+++ peripy/test/test_correction.py
@@ -20,7 +20,7 @@ def test_micromodulus_correction():
         ], dtype=np.intc)
     n_neigh = np.array([1, 2, 1], dtype=np.intc)
 
-    mmv_container = np.ones(np.shape(nl), dtype=np.float64)
+    mmv_container = np.ones(np.shape(nl), dtype=np.float32)
     actual_mmv = mmv_container.copy()
     horizon = 2.00
     set_micromodulus_function(actual_mmv, r0, nl, n_neigh, horizon, 0)
@@ -28,7 +28,7 @@ def test_micromodulus_correction():
         [0.5, 1],
         [0.5, 0.5],
         [0.5, 1]
-        ], dtype=np.float64)
+        ], dtype=np.float32)
     assert np.all(expected_mmv == actual_mmv)
     actual_mmv = mmv_container.copy()
     horizon = 1.00
@@ -37,7 +37,7 @@ def test_micromodulus_correction():
         [0.0, 1],
         [0.0, 0.0],
         [0.0, 1]
-        ], dtype=np.float64)
+        ], dtype=np.float32)
     assert np.all(expected_mmv == actual_mmv)
 
 
@@ -55,7 +55,7 @@ def test_volume_correction():
         [1, 0]
         ], dtype=np.intc)
     n_neigh = np.array([1, 2, 1], dtype=np.intc)
-    vlm_crtn_container = np.ones(np.shape(nl), dtype=np.float64)
+    vlm_crtn_container = np.ones(np.shape(nl), dtype=np.float32)
 
     actual_vlm_crtn = vlm_crtn_container.copy()
     horizon = 1.00
@@ -65,7 +65,7 @@ def test_volume_correction():
         [0.5, 1],
         [0.5, 0.5],
         [0.5, 1]
-        ], dtype=np.float64)
+        ], dtype=np.float32)
     assert np.allclose(actual_vlm_crtn, expected_vlm_crtn)
 
     actual_vlm_crtn = vlm_crtn_container.copy()
@@ -76,7 +76,7 @@ def test_volume_correction():
         [0.6, 1],
         [0.6, 0.6],
         [0.6, 1]
-        ], dtype=np.float64)
+        ], dtype=np.float32)
     assert np.allclose(actual_vlm_crtn, expected_vlm_crtn)
 
     actual_vlm_crtn = vlm_crtn_container.copy()
@@ -87,7 +87,7 @@ def test_volume_correction():
         [0.0, 1],
         [0.0, 0.0],
         [0.0, 1]
-        ], dtype=np.float64)
+        ], dtype=np.float32)
     assert np.allclose(actual_vlm_crtn, expected_vlm_crtn)
 
     actual_vlm_crtn = vlm_crtn_container.copy()
@@ -98,7 +98,7 @@ def test_volume_correction():
         [1.0, 1],
         [1.0, 1.0],
         [1.0, 1]
-        ], dtype=np.float64)
+        ], dtype=np.float32)
     expected_vlm_crtn = 1.0
     assert np.allclose(actual_vlm_crtn, expected_vlm_crtn)
 
@@ -114,9 +114,9 @@ class TestSurfaceCorrection():
             [1, 0]
             ], dtype=np.intc)
         n_neigh = np.array([1, 2, 1], dtype=np.intc)
-        stf_crtn_container = np.ones(np.shape(nl), dtype=np.float64)
-        average_volume = np.float64(1.0)
-        family_volume_bulk = np.float64(2.0)
+        stf_crtn_container = np.ones(np.shape(nl), dtype=np.float32)
+        average_volume = np.float32(1.0)
+        family_volume_bulk = np.float32(2.0)
         actual_stf_crtn = stf_crtn_container.copy()
         set_imprecise_surface_correction(actual_stf_crtn, nl, n_neigh,
                                          average_volume, family_volume_bulk)
@@ -124,20 +124,20 @@ class TestSurfaceCorrection():
             [4./3, 1],
             [4./3, 4./3],
             [4./3, 1]
-            ], dtype=np.float64)
+            ], dtype=np.float32)
         assert np.allclose(actual_stf_crtn, expected_stf_crtn)
 
     def test_precise(self):
         """Test surface corrections using precise nodal volumes."""
-        volume = np.array([1.0, 2.0, 1.0], dtype=np.float64)
+        volume = np.array([1.0, 2.0, 1.0], dtype=np.float32)
         nl = np.array([
             [1, 0],
             [0, 2],
             [1, 0]
             ], dtype=np.intc)
         n_neigh = np.array([1, 2, 1], dtype=np.intc)
-        stf_crtn_container = np.ones(np.shape(nl), dtype=np.float64)
-        family_volume_bulk = np.float64(2.0)
+        stf_crtn_container = np.ones(np.shape(nl), dtype=np.float32)
+        family_volume_bulk = np.float32(2.0)
         actual_stf_crtn = stf_crtn_container.copy()
         set_precise_surface_correction(actual_stf_crtn, nl, n_neigh,
                                        volume, family_volume_bulk)
@@ -145,5 +145,5 @@ class TestSurfaceCorrection():
             [4./4, 1],
             [4./4, 4./4],
             [4./4, 1]
-            ], dtype=np.float64)
+            ], dtype=np.float32)
         assert np.allclose(actual_stf_crtn, expected_stf_crtn)
